# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GraphDB API",
    "description": "GraphDB network service layer (HTTP/gRPC)",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/migration/stream/{space}/{label}": {
      "get": {
        "tags": [
          "Migration"
        ],
        "summary": "SSE endpoint to stream migration progress for a given space/label.",
        "description": "Example: GET /v1/migration/stream/{space}/{label}?is_edge=false",
        "operationId": "migration_progress_stream",
        "parameters": [
          {
            "name": "space",
            "in": "path",
            "description": "Space name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "label",
            "in": "path",
            "description": "Tag or edge type name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_edge",
            "in": "query",
            "description": "Whether the label is an edge type",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Migration progress event stream",
            "content": {
              "text/event-stream": {}
            }
          }
        }
      }
    },
...
    "/v1/query/stream": {
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Execute the query and stream the results",
        "operationId": "execute_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Query result event stream",
            "content": {
              "text/event-stream": {}
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/v1/statistics/sessions/{id}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Obtaining session statistics",
        "operationId": "session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session statistics",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StreamQueryRequest": {
        "type": "object",
        "description": "Streaming query request (SSE `/stream` endpoint).\n\nSingle-statement mode uses `query`. Batch-streaming mode uses\n`statements` (with shared `parameters` / `session_variables`, mirroring\n[`BatchQueryRequest`]) and emits one `statement_begin` /\n`statement_end` event pair per statement. The two modes are mutually\nexclusive: a non-empty `query` together with non-empty `statements` is\nrejected by the server.",
        "required": [
          "query",
          "session_id"
        ],
        "properties": {
          "event_buffer_capacity": {
            "type": "integer",
            "minimum": 0
          },
          "fail_fast": {
            "type": "boolean",
            "description": "Stop the batch at the first failing statement (default true).\nAlready produced statements are kept either way."
          },
          "parameters": {
            "type": "object",
            "description": "Query parameters bound to `@name` references in every statement\n(batch-streaming mode only).",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "query": {
            "type": "string"
          },
          "session_id": {
            "type": "integer",
            "format": "int64"
          },
          "session_variables": {
            "type": "object",
            "description": "Session variables bound to `$name` references in every statement\n(batch-streaming mode only). When omitted, the session-managed\nsnapshot is used, same as [`BatchQueryRequest`].",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "statements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Batch-streaming statements. Empty means single-statement mode."
          }
        }
      },
...
    }
  },
...
}